fix(park): a listening or unconnected socket forwards instead of parking — A4's can_block rule, which ADR-0037 §6 said to move before deleting the hooks and I did not; gate now compares per-test sets, not just counts (V-50)
- park.rs would_block(): SO_ACCEPTCONN -> never park a data call; connected -> park; unconnected -> only a bound SOCK_DGRAM parks (recvfrom waits legitimately), else ENOTCONN reaches the caller
- socket_read_params.phpt: PASSED at 6780e1b, timed out from 17a2ceb, passes again; fiber sockets 85 (high-water), fiber streams 125
- scripts/ci-gate.sh check_set(): a test that PASSED in HEAD's committed .tsv and does not now is a regression; a swap (one test breaks, another heals) is invisible to a total, which is how this shipped. Warns under $CI until a CI-box set is committed
- build.rs: never interpose fcntl (opcache's zend_shared_alloc_lock holds the TSRM mutex across F_SETLKW - research 30 (d))
- ADR-0024: blocking calls on regular files are a non-goal (session flock stalls a thread - BACKLOG R-SESS)
- ADR-0020: getaddrinfo corrected per research 31 - no fd means no park; forward the real call to a blocking pool behind Op::Custom, fabricate nothing, freeaddrinfo needs no interposer
- agent research: 30 groups (d)/(e), 31 DNS, 32 self-check + blocked-in-fiber detector spec, 33 phpt failure classification; pain-map/GOALS/ROADMAP reconciled
- H36 shim written (bench/e18/locklib.c + driver + bench script); the three internal functions it needs are BACKLOG H36

+++ bench/php/e18_selfcheck.php
@@ -46,20 +46,20 @@ $fs[] = Ignis\async(static function () use ($ms): void {
     [$a, $b] = $pair;
     socket_set_option($a, SOL_SOCKET, SO_RCVTIMEO, ['sec' => intdiv($ms, 1000), 'usec' => ($ms % 1000) * 1000]);
     $r = socket_recv($a, $buf, 16, 0);
-    $elapsed = intdiv(hrtime(true) - $t0, 1_000_000);
+    $elapsed = intdiv((int) hrtime(true) - $t0, 1_000_000);
     printf("e18_selfcheck: case=blocks(recv, off policy) result=%s elapsed_ms=%d\n", var_export($r, true), $elapsed);
 });
 
 $fs[] = Ignis\async(static function () use ($ms): void {
     $t0 = hrtime(true);
     usleep($ms * 1000);
-    $elapsed = intdiv(hrtime(true) - $t0, 1_000_000);
+    $elapsed = intdiv((int) hrtime(true) - $t0, 1_000_000);
     printf("e18_selfcheck: case=parks(usleep, on policy) elapsed_ms=%d\n", $elapsed);
 });
 
 $t0 = hrtime(true);
 Ignis\all($fs);
-$total = intdiv(hrtime(true) - $t0, 1_000_000);
+$total = intdiv((int) hrtime(true) - $t0, 1_000_000);
 printf(
     "e18_selfcheck: total_wall_ms=%d (serialized ~= %d if 'blocks' really blocked the thread; ~= %d if both ran concurrently)\n",
     $total,
